子游戏连接网关整理

+++ ugk-game/game-galactic-kittens/galactic-kittens-game/Assets/Scripts/Game/Handlers/ServerHandler.cs
@@ -2,6 +2,7 @@
 using Game.Manager;
 using Google.Protobuf;
 using Common.Network;
+using kcp2k;
 using UnityEngine;
 
 
@@ -15,7 +16,7 @@ namespace Game.Handlers
         /// <summary>
         /// 心跳
         /// </summary>
-        [MessageMap(MID.ServerHeartRes)]
+        [MessageMap((int)MID.ServerHeartRes)]
         private static void Heart(Player player, UgkMessage ugkMessage)
         {
             var response = new ServerHeartResponse();
